requirements.txt
Django==4.2.3
django-crispy-forms==1.14.0
pyrebase4